requirements.txt
networkx>=3.0
kuzu>=0.6.0
openai>=1.0
transformers>=4.40
peft>=0.10
datasets>=2.0
accelerate>=0.30
bitsandbytes
pandas>=2.0
pydantic>=2.0
typer>=0.12
rich>=13.0
torch>=2.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.5.0